backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
opencv-python-headless
pillow
python-dotenv
pydantic
